Avoid mypyc errors
